Backend/python/requirements.txt
# Microsoft Agent Framework - Core and integrations
agent-framework --pre

# Web framework
fastapi[all]
uvicorn[standard]

# Configuration and environment
python-dotenv
pyyaml

# Azure integrations
azure-identity
azure-ai-inference
azure-ai-contentsafety
azure-ai-projects>=1.0.0b4
azure-ai-agents>=1.0.0b1
openai

# Data handling
pydantic
pydantic-settings

# AWS integration
boto3
botocore

# Google AI integration  
google-generativeai

# Utility libraries
aiofiles
httpx

# Logging and monitoring
structlog

# Session storage options
redis

# Development dependencies
pytest
pytest-asyncio
black
isort
mypy

# Observability and Telemetry Packages for Application Insights Integration
# Add these to your requirements.txt

# OpenTelemetry Core
opentelemetry-api>=1.24.0
opentelemetry-sdk>=1.24.0

# OpenTelemetry Instrumentation
opentelemetry-instrumentation>=0.45b0
opentelemetry-instrumentation-requests>=0.45b0
opentelemetry-instrumentation-fastapi>=0.45b0

# Azure Monitor Exporter for Application Insights
azure-monitor-opentelemetry-exporter>=1.0.0b27

# Note: These packages enable:
# - Distributed tracing with OpenTelemetry
# - Automatic HTTP request instrumentation
# - FastAPI endpoint tracing
# - Azure Monitor (Application Insights) export
# - Console export for local debugging